Chatroom msg

+++ studybud/base/views.py
@@ -1,12 +1,15 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 
 
-from .models import Room, Topic
+
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 
@@ -22,6 +25,11 @@ from django.http import HttpResponse
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -41,12 +49,34 @@ def loginPage(request):
             messages.error(request, 'Username OR password does not exist')
 
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            #some customization
+            user= form.save(commit=False)
+            user.username= user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        
+        else:
+            messages.error(request, 'An error occured during registrations')
+            
+           
+
+    return render(request, 'base/login_register.html', {'form': form})
+ 
+
 
 def home(request):
 
@@ -67,12 +97,22 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created_at')
+    
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        return redirect('room', pk=room.id)
+    
+    context = {'room':room, 'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
 
 
-
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -85,10 +125,15 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id =pk)
     form = RoomForm(instance=room)
+
+
+    # Restricting anyone to update with id
+    if request.user != room.host:
+        return HttpResponse('Youre not allowed here')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -99,9 +144,15 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+
+     # Restricting anyone to delete with id
+    if request.user != room.host:
+        return HttpResponse('Youre not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
